more EF core model updates

+++ StreamerBotLib/DataSQL/Models/OverlayServices.cs
@@ -7,12 +7,20 @@ using System.ComponentModel.DataAnnotations.Schema;
 namespace StreamerBotLib.DataSQL.Models
 {
     [Index(nameof(Id), nameof(OverlayType), nameof(OverlayAction))]
-    public class OverlayServices(uint id = 0, bool isEnabled = false, decimal duration = 0, OverlayTypes overlayType = default, string overlayAction = null, string userName = null, bool useChatMsg = false, string message = null, string imageFile = null)
+    public class OverlayServices(uint id = 0,
+                                 bool isEnabled = false,
+                                 ushort duration = 0,
+                                 OverlayTypes overlayType = default,
+                                 string overlayAction = null,
+                                 string userName = null,
+                                 bool useChatMsg = false,
+                                 string message = null,
+                                 string imageFile = null)
     {
         [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
         public uint Id { get; set; } = id;
         public bool IsEnabled { get; set; } = isEnabled;
-        public decimal Duration { get; set; } = duration;
+        public ushort Duration { get; set; } = duration;
         public OverlayTypes OverlayType { get; set; } = overlayType;
         public string OverlayAction { get; set; } = overlayAction;
         public string UserName { get; set; } = userName;
